src/backend/pyproject.toml
[project]
name = "linkinthe-video"
version = "0.1.0"
description = "Backend for linkinthe.video"
readme = "../../README.md"
authors = [{ name = "linkinthe.video" }]
requires-python = ">=3.11"
dependencies = [
    "Django>=5.1",
    "django-shinobi>=0.4.0",
    "django-allauth>=65.0",
    "django-environ>=0.11",
    "django-anymail[resend]>=12.0",
    "PyJWT>=2.9",
    "cryptography>=43.0.1",
    "yt-dlp>=2024.8.6",
    "ffmpeg-python>=0.2.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.7",
    "mypy>=1.13",
    "django-stubs>=5.1.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 79
src = ["."]
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",
    "F",
    "I",
    "D",
    "W",
    "N",
    "ASYNC",
    "B",
]
ignore = ["D100", "D104"]
extend-ignore = ["D"]

[tool.ruff.lint.per-file-ignores]
"**/migrations/*.py" = ["D", "E501"]
"src/backend/**/tests.py" = ["D"]
"src/backend/pipeline/**/*.py" = ["D"]
"src/backend/user/api.py" = ["D"]
"src/backend/video/api.py" = ["D"]
"src/backend/user/admin.py" = ["D"]
"src/backend/video/admin.py" = ["D"]
"src/backend/user/apps.py" = ["D"]
"src/backend/video/apps.py" = ["D"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"

[tool.mypy]
python_version = "3.11"
plugins = ["mypy_django_plugin.main"]
disallow_untyped_defs = true
strict_equality = true
check_untyped_defs = true
ignore_missing_imports = false
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true
mypy_path = "."

[[tool.mypy.overrides]]
module = ["environ"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["allauth.*", "ninja.*", "ffmpeg.*", "yt_dlp.*", "litellm.*", "modal.*"]
ignore_missing_imports = true

[tool.django-stubs]
django_settings_module = "config.settings"

[dependency-groups]
dev = ["ruff>=0.7", "mypy>=1.13", "django-stubs>=5.1.1"]

[tool.hatch.build.targets.wheel]
packages = ["config"]
